--story=消息、评论、收藏怎删改查已经OK

+++ server/sql/collectSql.js
@@ -1,19 +1,19 @@
 const {sqlQuery} = require('./sql.js')
 module.exports = {
   getCollects (userId, offset = 0, limit = 20) {
-    const sql = `select * from collect_table where user_id=${userId} order offset ${offset} limit ${limit} offset  by create_time desc;`
-    return sqlQuery(sql).then(data => data || [])
+    const sql = `select * from collect_table where user_id=${userId} order by create_time desc limit ${limit} offset ${offset};`
+    return sqlQuery(sql).then(data => data)
   },
   getCollectsTotal (userId) {
-    const sql = `select count(*) from collect_table  where user_id=${userId};`
+    const sql = `select count(*) as count from collect_table  where user_id=${userId};`
     return sqlQuery(sql).then(data => data && data[0])
   },
   addCollect ({messageId, userId}) {
     const sql = `insert into collect_table (message_id, user_id) values (${messageId}, ${userId});`
-    return sqlQuery(sql).then(data => data[0]).then(data => data && data[0])
+    return sqlQuery(sql).then(data => data)
   },
-  delCollect (collectId) {
-    const sql = `delete from collect_table where collect_id = ${collectId};`
-    return sqlQuery(sql).then(data => data[0])
+  delCollect ({collectId, userId}) {
+    const sql = `delete from collect_table where collect_id = ${collectId} and user_id=${userId};`
+    return sqlQuery(sql).then(data => data)
   }
 }
